Moved analysis_modules.py into bilayer_analyzer and renamed to canned_analysis_protocols.py. Updated the names and import statements of the corresponding tests.
