--- pybilt/analysis_modules/__init__.py
@@ -1,11 +1,0 @@
-"""
-The analysis_module module contains functions that perform sets of related
-or interdependent analyses using tools from with pybilt; this includes
- outputting data and generating plots when relevant. These are meant to
-streamline the use of 'common' pybilt analyses by removing the need for the
- user to directly construct specific anaysis objects, e.g. using the BilayerAnalysis
- object.
-"""
-
-from pybilt.analysis_modules.analysis_modules import *
-
