+++ pybilt/bilayer_analyzer/__init__.py
@@ -1,1 +1,4 @@
 from bilayer_analyzer import BilayerAnalyzer
+import canned_analysis_protocols
+from bilayer_analyzer import print_valid_analyses, print_analysis_settings
+from bilayer_analyzer import print_available_plots
