+++ tests/test_ba_canned_protocol_PN_orientation_angle.py
@@ -1,5 +1,5 @@
 
-from pybilt.analysis_modules import PN_orientational_angle
+from from pybilt.bilayer_analyzer.canned_analysis_protocols import PN_orientational_angle
 
 
 def test_analysis_module_PN_orientation_anlge():
