+++ tests/test_ba_canned_protocol_area_per_lipid.py
@@ -1,5 +1,4 @@
-from pybilt.analysis_modules import area_per_lipid
-
+from pybilt.bilayer_analyzer.canned_analysis_protocols import area_per_lipid
 
 def test_analysis_module_area_per_lipid():
     sel_string = "not resname CLA and not resname TIP3 and not resname POT"
