+++ tests/test_ba_canned_protocol_compressibility.py
@@ -1,5 +1,5 @@
 
-from pybilt.analysis_modules import compressibility
+from pybilt.bilayer_analyzer.canned_analysis_protocols import compressibility
 
 
 def test_analysis_module_compressibility():
