+++ tests/test_ba_canned_protocol_displacement_lipid_type_correlation.py
@@ -1,4 +1,4 @@
-from pybilt.analysis_modules import normal_displacement_lipid_type_correlation
+from pybilt.bilayer_analyzer.canned_analysis_protocols import normal_displacement_lipid_type_correlation
 
 
 def test_analysis_module_normal_displacement_lipid_type_correlation():
