+++ tests/test_ba_canned_protocol_lipid_grid_maps.py
@@ -1,4 +1,4 @@
-from pybilt.analysis_modules import lipid_grid_maps
+from pybilt.bilayer_analyzer.canned_analysis_protocols import lipid_grid_maps
 
 
 def test_analysis_module_lipid_grid_maps():
