+++ tests/test_ba_canned_protocol_position_density_map_2d.py
@@ -1,4 +1,4 @@
-from pybilt.analysis_modules import position_density_maps_2d
+from pybilt.bilayer_analyzer.canned_analysis_protocols import position_density_maps_2d
 
 
 def test_analysis_module_position_density_maps_2d():
